+++ src/main/java/com/example/ubergo/controller/RideDistributionController.java
@@ -1,6 +1,6 @@
 package com.example.ubergo.controller;
 
-import com.example.ubergo.DTO.*;
+import com.example.ubergo.DTO.RestDTO.*;
 import com.example.ubergo.service.RideDistributionService;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,16 +18,16 @@ public class RideDistributionController {
     }
 
     @RequestMapping(value = "/ride/{rid}/takeOrder",method = RequestMethod.PUT)
-    public GeneralMessageDTO takeOrder(@PathVariable int rid, @RequestBody TakeOrderReqDTO takeOrderReqDTO){
+    public GeneralMessageDTO takeOrder(@PathVariable long rid, @RequestBody TakeOrderReqDTO takeOrderReqDTO){
         return rideDistributionService.takeOrder(rid,takeOrderReqDTO);
     }
     @RequestMapping(value = "/ride/{rid}/cancel",method = RequestMethod.PUT)
-    public GeneralMessageDTO cancelRide(@PathVariable int rid, @RequestBody CancelRideReqDTO cancelRideReqDTO){
+    public GeneralMessageDTO cancelRide(@PathVariable long rid, @RequestBody CancelRideReqDTO cancelRideReqDTO){
         return rideDistributionService.cancelRide(rid,cancelRideReqDTO);
     }
 
     @RequestMapping(value = "/ride/{rid}",method = RequestMethod.GET)
-    public GeneralMessageDTO getRideInfo(@PathVariable int rid, @RequestParam(value = "long") Double longitide, @RequestParam(value = "lat") Double latitude){
+    public GeneralMessageDTO getRideInfo(@PathVariable long rid, @RequestParam(value = "long") Double longitide, @RequestParam(value = "lat") Double latitude){
         //I haven't made use of location given
         LocationDTO locationDTO=new LocationDTO(longitide,latitude);
         return rideDistributionService.getRiderInfo(rid,locationDTO);
