+++ src/main/java/com/example/ubergo/entity/User.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class User {
-    private int uid;
+    private Long uid;
     private String mobileNumber;
     private String identity;
     private String password;
     private String userName;
     private String licensePlateNumber;
     private String carType;
-    private float totalTripLength;
+    private Double totalTripLength;
     private String regionalProvince;
     private String city;
